+++ src/runOnPullRequest.js
@@ -2,44 +2,25 @@
 
 import {type Octokit$IssuesListCommentsResponseItem} from '@octokit/rest';
 
-type CommentHeaders = 'Reviewers:\n' | 'Required reviewers:\n' | 'Notified:\n';
-
 import {
     getReviewers,
     getNotified,
     getFileDiffs,
     parseExistingComments,
     getFilteredLists,
+    makeCommentBody,
 } from './utils';
 import {execCmd} from './execCmd';
-import {ownerAndRepo, context, extraPermGithub, type Context} from './setup';
+import {ownerAndRepo, context, extraPermGithub} from './setup';
 import {
     GERALD_COMMENT_FOOTER,
     PULL_REQUEST,
-    PUSH,
     GERALD_COMMENT_HEADER,
     GERALD_COMMENT_NOTIFIED_HEADER,
     GERALD_COMMENT_REQ_REVIEWERS_HEADER,
     GERALD_COMMENT_REVIEWERS_HEADER,
     MATCH_COMMENT_HEADER_REGEX,
-    GERALD_COMMIT_COMMENT_HEADER,
 } from './constants';
-
-const makeCommentBody = (
-    peopleToFiles: {[string]: Array<string>, ...},
-    sectionHeader: CommentHeaders,
-) => {
-    const names: string[] = Object.keys(peopleToFiles);
-    if (names.length) {
-        let body = `### ${sectionHeader}`;
-        names.forEach((person: string) => {
-            const files = peopleToFiles[person];
-            body += `${person} for changes to \`${files.join('`, `')}\`\n\n`;
-        });
-        return body;
-    }
-    return '';
-};
 
 /**
  * @desc Helper function to update, delete, or create a comment
@@ -81,27 +62,7 @@ const updatePullRequestComment = async (
     }
 };
 
-const makeCommitComment = async (
-    peopleToFiles: {[string]: Array<string>, ...},
-    commitSHA: string,
-) => {
-    const names: string[] = Object.keys(peopleToFiles);
-    if (peopleToFiles && names.length) {
-        let body: string = GERALD_COMMIT_COMMENT_HEADER;
-        names.forEach((person: string) => {
-            const files = peopleToFiles[person];
-            body += `${person} for changes to \`${files.join('`, `')}\`\n`;
-        });
-
-        await extraPermGithub.repos.createCommitComment({
-            ...ownerAndRepo,
-            commit_sha: commitSHA,
-            body: body,
-        });
-    }
-};
-
-export const runPullRequest = async () => {
+export const runOnPullRequest = async () => {
     // get the files changed between the head of this branch and the origin of the base branch
     const filesChanged = (
         await execCmd('git', [
@@ -148,47 +109,3 @@ export const runPullRequest = async () => {
 
     await updatePullRequestComment(megaComment, notified, reviewers, requiredReviewers);
 };
-
-export const runPush = async (usedContext: Context) => {
-    // loop through each commit in the push
-    let prevCommit = usedContext.payload.before;
-    for (const commit of usedContext.payload.commits) {
-        const commitData = await extraPermGithub.git.getCommit({
-            ...ownerAndRepo,
-            commit_sha: commit.id,
-        });
-
-        // commits with >1 parent are merge commits. we want to ignore those
-        if (commitData.data.parents.length === 1) {
-            const filesChanged = (
-                await execCmd('git', [
-                    'diff',
-                    `${prevCommit}...${commitData.data.sha}`,
-                    '--name-only',
-                ])
-            ).split('\n');
-            const fileDiffs = await getFileDiffs(`${prevCommit}...${commitData.data.sha}`);
-            const notified = getNotified(filesChanged, fileDiffs, PUSH);
-
-            await makeCommitComment(notified, commitData.data.sha);
-        }
-
-        // we also want to ignore the diff of a merge commit
-        prevCommit = commitData.data.sha;
-    }
-};
-
-export type __TestCommit = {
-    author: '__testAuthor',
-    comment_count: -1,
-    committer: '__testCommitter',
-    id: string,
-    message: string,
-    tree: '__TESTING__',
-    url: '__TESTING__',
-    verification: '__TESTING__',
-};
-
-export const __makeCommitComment = makeCommitComment;
-export const __makeCommentBody = makeCommentBody;
-export const __extraPermGithub = extraPermGithub;
